stat/requirements.txt
Flask
requests
confluent-kafka
clickhouse-driver
clickhouse_connect
grpcio
grpcio-tools